Make sure __u64 is defined even for -std=c99
The x86 asm headers define __u64 regardless of __STRICT_ANSI__.
The linux/videodev2.h header requires __u64 to be defined, thus
this fixes compiling with -std=c99 when including the
linux/videodev2.h header.

In glibc, the asm/types.h header defines __u64 regardless of
__STRICT_ANSI__. The original kernel header needs to be adjusted
in this case, since it is used as source for the bionic userland
headers.

Change-Id: I07ecec2774049eb7e8e83f4d5b6af326b26a8761